gridcapa-swe-runner-api/src/test/java/com/farao_community/farao/swe/api/resource/SweRequestTest.java
/*
 * Copyright (c) 2022, RTE (http://www.rte-france.com)
 *  This Source Code Form is subject to the terms of the Mozilla Public
 *  License, v. 2.0. If a copy of the MPL was not distributed with this
 *  file, You can obtain one at http://mozilla.org/MPL/2.0/.
 */
package com.farao_community.farao.swe.api.resource;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;

/**
 * @author Theo Pascoli {@literal <theo.pascoli at rte-france.com>}
 */
public class SweRequestTest {

    @Test
    void simpleRequestTest() {
        SweRequest sweRequest = new SweRequest("id");
        assertNotNull(sweRequest);
        assertEquals("id", sweRequest.getId());
    }

}
